Move action def to dir
