+++ ui/src/store/action/snakeActions.ts
@@ -1,5 +1,5 @@
-import {ActionUnion, createAction, createActionWithPayload} from '../typeHelper';
-import {ActionTypes} from '../actions';
+import {ActionUnion, createAction, createActionWithPayload} from './typeHelper';
+import {ActionTypes} from './actions';
 import {Direction, Point} from '../../domain';
 
 const biteSelf = createAction(ActionTypes.SNAKE_BITE_SELF);
