+++ ui/src/store/action/typeHelper.ts
@@ -1,7 +1,7 @@
 import {Action} from 'redux';
-import {checkNonEmpty} from '../utils';
+import {checkNonEmpty} from '../../utils';
 
-export interface ActionWithPayload<T extends string, P> extends Action<T> {
+interface ActionWithPayload<T extends string, P> extends Action<T> {
     readonly payload: P;
 }
