+++ ui/src/store/games/index.ts
@@ -1,7 +1,7 @@
 import {snakeEpic} from './snakeEpic';
 import {combineEpics} from 'redux-observable';
 
-export {SnakeActions, SnakeAction} from './snakeActions';
+export {SnakeActions, SnakeAction} from '../action/snakeActions';
 export {snakeGameReducer} from './snakeReducer';
 export {SnakeGameState, DefaultSnakeGameState} from './snakeState';
