+++ ui/src/store/sound/coreAudioEpic.ts
@@ -1,7 +1,7 @@
 import {Observable} from 'rxjs';
 import {AppAction, AppState} from '../index';
 import {combineEpics, ofType, StateObservable} from 'redux-observable';
-import {ActionTypes} from '../actions';
+import {ActionTypes} from '../action';
 import {filter, tap, withLatestFrom} from 'rxjs/operators';
 import {SoundEffects} from './types';
 import {SystemStatus} from '../../domain';
@@ -16,6 +16,16 @@ const menuAudioEpic = (action$: Observable<AppAction>, state$: StateObservable<A
     );
 };
 
+const startAudioEpic = (action$: Observable<AppAction>, state$: StateObservable<AppState>) => {
+    return action$.pipe(
+        ofType(ActionTypes.START),
+        withLatestFrom(state$),
+        filter(([, s]) => s.core.status === SystemStatus.MENU),
+        tap(() => SoundEffects.playSfxCoreMenu()),
+        filter(() => false),
+    );
+};
+
 export const coreAudioEpic = {
-    epic: combineEpics(menuAudioEpic),
+    epic: combineEpics(menuAudioEpic, startAudioEpic),
 };
